# server/api/openapi-with-docs.json
{
  "x-generator": "NSwag v14.7.1.0 (NJsonSchema v11.6.1.0 (Newtonsoft.Json v13.0.0.0))",
  "openapi": "3.0.0",
  "info": {
    "title": "Belman API",
    "version": "v1"
  },
  "paths": {
    "/api/alerts": {
      "get": {
        "tags": [
          "Alerts"
        ],
        "operationId": "Alerts_GetAlerts",
        "parameters": [
          {
            "name": "unresolvedOnly",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": true
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlertDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/alerts/{id}/resolve": {
      "patch": {
        "tags": [
          "Alerts"
        ],
        "operationId": "Alerts_ResolveAlert",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/categories": {
      "get": {
        "tags": [
          "Categories"
        ],
        "operationId": "Categories_GetCategories",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/dispatches": {
      "get": {
        "tags": [
          "Dispatches"
        ],
        "operationId": "Dispatches_GetDispatches",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DispatchDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/dispatches/{id}/mark-ready": {
      "patch": {
        "tags": [
          "Dispatches"
        ],
        "operationId": "Dispatches_MarkReady",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/products": {
      "get": {
        "tags": [
          "Products"
        ],
        "operationId": "Products_GetProducts",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "categoryId",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 3
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 4
          },
          {
            "name": "material",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 5
          },
          {
            "name": "pressure",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 6
          },
          {
            "name": "inStockOnly",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "x-position": 7
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/products/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "operationId": "Products_GetProductById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/stats": {
      "get": {
        "tags": [
          "Stats"
        ],
        "operationId": "Stats_GetStats",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardStatsDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/warehouse": {
      "get": {
        "tags": [
          "Warehouse"
        ],
        "operationId": "Warehouse_GetWarehouseItems",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WarehouseItemDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/warehouse/{productId}/pick": {
      "post": {
        "tags": [
          "Warehouse"
        ],
        "operationId": "Warehouse_PickProduct",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PickRequestDto"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PickResultDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AlertDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "productId": {
            "type": "string"
          },
          "productReference": {
            "type": "string"
          },
          "productName": {
            "type": "string"
          },
          "alertType": {
            "type": "string"
          },
          "severity": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "stockQuantity": {
            "type": "integer",
            "format": "int32"
          },
          "maxStock": {
            "type": "integer",
            "format": "int32"
          },
          "stockPercentage": {
            "type": "number",
            "format": "decimal"
          },
          "isResolved": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "resolvedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "CategoryDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "productCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "DispatchDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "productId": {
            "type": "string"
          },
          "productReference": {
            "type": "string"
          },
          "productName": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "requestedBy": {
            "type": "string"
          },
          "workshop": {
            "type": "string"
          },
          "requestedAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "ProductDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "reference": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "categoryId": {
            "type": "string"
          },
          "categoryName": {
            "type": "string"
          },
          "size": {
            "type": "string"
          },
          "material": {
            "type": "string"
          },
          "pressure": {
            "type": "string"
          },
          "temperature": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "zone": {
            "type": "string"
          },
          "shelf": {
            "type": "string"
          },
          "bin": {
            "type": "string"
          },
          "stockQuantity": {
            "type": "integer",
            "format": "int32"
          },
          "maxStock": {
            "type": "integer",
            "format": "int32"
          },
          "stockPercentage": {
            "type": "number",
            "format": "decimal"
          },
          "isLowInventory": {
            "type": "boolean"
          }
        }
      },
      "DashboardStatsDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "dbSync": {
            "type": "string"
          },
          "totalItems": {
            "type": "integer",
            "format": "int32"
          },
          "pendingRequests": {
            "type": "integer",
            "format": "int32"
          },
          "activeAlerts": {
            "type": "integer",
            "format": "int32"
          },
          "workshop": {
            "type": "string"
          }
        }
      },
      "WarehouseItemDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "reference": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "categoryName": {
            "type": "string"
          },
          "size": {
            "type": "string"
          },
          "pressure": {
            "type": "string"
          },
          "zone": {
            "type": "string"
          },
          "shelf": {
            "type": "string"
          },
          "bin": {
            "type": "string"
          },
          "stockQuantity": {
            "type": "integer",
            "format": "int32"
          },
          "maxStock": {
            "type": "integer",
            "format": "int32"
          },
          "stockPercentage": {
            "type": "number",
            "format": "decimal"
          },
          "isLowInventory": {
            "type": "boolean"
          }
        }
      },
      "PickResultDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "item": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/WarehouseItemDto"
              }
            ]
          }
        }
      },
      "PickRequestDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "quantity": {
            "type": "integer",
            "format": "int32"
          }
        }
      }
    }
  }
}